Minor changes.
Implemented PUD , changed timer_sleep to use longs. Fixed Makefile.

+++ timer.c
@@ -1,7 +1,7 @@
 #include "timer.h"
 extern void miniWait ( unsigned int );
-void timer_sleep(unsigned int time) {
-  unsigned int* timeLower = TIMER_COUNTER_LOWER;
-  unsigned int waitUntil = *timeLower + time;
+void timer_sleep(unsigned long time) {
+  unsigned long* timeLower = TIMER_COUNTER_LOWER;
+  unsigned long waitUntil = *timeLower + time;
   while(*timeLower < waitUntil) miniWait(0);
 }
